fix(gamma-wl): make outputs a linked list for pointer stability
Before this commit, the list of outputs was kept in a dynamic array
that is resized using realloc when an output is added or removed.

This is problematic because pointers to the elements of the dynamic
array are passed to `zwlr_gamma_control_v1_add_listener` so that they
become the `data` argument of the listener's callback. When an output
is added/removed, the dynamic array may get relocated, thus pointers
to other outputs become invalid/dangling, and gammastep may crash.

This can be reproduced quite consistently by adding a `sleep` before
the call to `wl_display_roundtrip` in `wayland_set_temperature`, and
plugging or unplugging an output during the `sleep`.

This commit fixes it by making `outputs` a wl_list, so outputs can be
added or removed without invalidating pointers to other outputs.